+++ tokenize.c
@@ -110,10 +110,10 @@ void expect(char *s) {
 
 // 次のトークンが数値の場合、トークンを一つ読み進めてその数値を返す。
 // それ以外の場合にはエラーを報告する
-int expect_number() {
+long expect_number() {
   if (token->kind != TK_NUM)
     error_tok(token, "数ではありません");
-  int val = token->val;
+  long val = token->val;
   token = token->next;
   return val;
 }
